Correctly delete/transfer users and ther projects in one go

+++ src/Data/Services/UserManagerDeletionExtension.cs
@@ -9,37 +9,42 @@ public static class UserManagerDeletionExtension
     extension(UserManager<ApplicationUser> self)
     {
         /// <summary>
-        /// Deletes a user’s projects, accesses and then the user itself.
+        /// Deletes a <paramref name="user"/> and then their projects, or, if <paramref name="transferRecipient"/> is set, transfers them.
         /// </summary>
-        public async Task<IdentityResult> DeleteUserAndProjectsAsync(ApplicationUser user)
+        public async Task<IdentityResult> DeleteUserAndProjectsAsync(ApplicationUser user, ApplicationUser? transferRecipient = null)
         {
             ProjectService projectService = self.ServiceProvider.GetRequiredService<ProjectService>();
             AccessControlService accessControlService = self.ServiceProvider.GetRequiredService<AccessControlService>();
 
             List<Project> projects = await projectService.GetOwnedProjectsAsync(user);
 
-            try
-            {
-                projects.ForEach(async p => await projectService.Delete(p));
-            }
-            catch (Exception ex)
-            {
-                return IdentityResult.Failed([new IdentityError() { Description = ex.Message }]);
-            }
-
-            if (!await accessControlService.DeleteUserAccessesAsync(user))
-            {
-                return IdentityResult.Failed([new IdentityError() { Description = "Failed to delete all accesses of user." }]);
-            }
-
-            // Prevents “Optimistic concurrency failure: Object has been modified”
             IdentityResult result = await self.DeleteAsync(user);
             if (!result.Succeeded)
             {
                 return result;
             }
 
-            // await accessControlService.DeleteOrphansAsync();
+            await accessControlService.DeleteOrphansAsync();
+
+            if (transferRecipient is not null)
+            {
+                result = await TransferNewlyOrphanedProjects(self, projects, targetUser: transferRecipient);
+                if (!result.Succeeded)
+                {
+                    return result;
+                }
+            }
+            else
+            {
+                try
+                {
+                    projects.ForEach(async p => await projectService.Delete(p));
+                }
+                catch (Exception ex)
+                {
+                    return IdentityResult.Failed([new IdentityError() { Description = ex.Message }]);
+                }
+            }
 
             return result;
         }
@@ -64,6 +69,44 @@ public static class UserManagerDeletionExtension
                     try
                     {
                         if (!await projectService.TransferOwnershipAsync(p.Id, targetUser, doNotVerifyOwnership))
+                            throw new Exception($"Failed to transfer project {p.ProjectName}");
+                    }
+                    catch (Exception ex)
+                    {
+                        oex.Add(ex);
+                    }
+                });
+                if (oex is not [])
+                {
+                    throw new Exception(string.Join('\n', oex.Select(e => e.Message)));
+                }
+            }
+            catch (Exception ex)
+            {
+                return IdentityResult.Failed([new IdentityError() { Description = ex.Message }]);
+            }
+
+            return IdentityResult.Success;
+        }
+
+        /// <summary>
+        /// Transfers a list of projects to the <paramref name="targetUser"/>.
+        /// </summery>
+        private async Task<IdentityResult> TransferNewlyOrphanedProjects
+        (
+            List<Project> projects, ApplicationUser targetUser
+        )
+        {
+            ProjectService projectService = self.ServiceProvider.GetRequiredService<ProjectService>();
+
+            try
+            {
+                List<Exception> oex = [];
+                projects.ForEach(async p =>
+                {
+                    try
+                    {
+                        if (!await projectService.TransferOrphanedProjectAsync(p.Id, targetUser))
                             throw new Exception($"Failed to transfer project {p.ProjectName}");
                     }
                     catch (Exception ex)
